+++ castero/__init__.py
@@ -31,10 +31,10 @@ Commands
     enter        - play selected feed/episode
     space        - add selected feed/episode to queue
     c            - clear the queue
-    p            - pause/play the current episode
     n            - go to the next episode in the queue
-    f            - seek forward
-    b            - seek backward
+    p or k       - pause/play the current episode
+    f or l       - seek forward
+    b or j       - seek backward
 """ % (
     __title__,
     __version__,